requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coincurve==15.0.1
coloredlogs==15.0.1
cryptography==3.3.2
cycler==0.11.0
cytoolz==0.11.2
Deprecated==1.2.13
eciespy==0.3.11
enforce-typing==1.0.0.post1
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.4.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fonttools==4.34.4
frozenlist==1.3.0
hexbytes==0.2.2
humanfriendly==10.0
idna==3.3
ipfshttpclient==0.8.0a2
json-sempai==0.4.0
jsonschema==3.2.0
kiwisolver==1.4.4
lru-dict==1.1.8
matplotlib==3.5.2
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.23.1
ocean-contracts==1.1.7
ocean-lib==1.6.1
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.20.1
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyJWT==2.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
rlp==2.0.1
scipy==1.8.1
six==1.16.0
tomli==2.0.1
toolz==0.11.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.10
varint==1.0.2
web3==5.28.0
websockets==9.1
wrapt==1.14.1
yarl==1.7.2
